+++ src/lib/constants.ts
@@ -1,0 +1,5 @@
+import type { Subscriber } from "./types";
+
+export const prefix = "/*framebus*/";
+export const childWindows: Window[] = [];
+export const subscribers: Subscriber = {};